Move EOS infos into a folder
