+++ src/eos/eosInfos.ts
@@ -1,0 +1,9 @@
+import { eos } from './info/eosInfo'
+import { telos } from './info/telosInfo'
+import { wax } from './info/waxInfo'
+
+export const eosPlugins = {
+  eos,
+  telos,
+  wax
+}
