--- src/eos/index.ts
@@ -1,9 +1,0 @@
-import { eos } from './eosInfo'
-import { telos } from './telosInfo'
-import { wax } from './waxInfo'
-
-export const eosPlugins = {
-  eos,
-  telos,
-  wax
-}
