+++ src/eos/info/waxInfo.ts
@@ -1,8 +1,8 @@
 import { EdgeCurrencyInfo } from 'edge-core-js/types'
 
-import { makeOuterPlugin } from '../common/innerPlugin'
-import type { EosTools } from './eosPlugin'
-import { EosNetworkInfo, eosOtherMethodNames } from './eosTypes'
+import { makeOuterPlugin } from '../../common/innerPlugin'
+import type { EosTools } from '../eosPlugin'
+import { EosNetworkInfo, eosOtherMethodNames } from '../eosTypes'
 
 // ----WAX MAIN NET----
 export const waxNetworkInfo: EosNetworkInfo = {
@@ -52,6 +52,6 @@ export const wax = makeOuterPlugin<EosNetworkInfo, EosTools>({
   otherMethodNames: eosOtherMethodNames,
 
   async getInnerPlugin() {
-    return await import('./eosPlugin')
+    return await import('../eosPlugin')
   }
 })
